move scripts/magic_folder_cli to magic_folder/cli
