+++ src/magic_folder/cli.py
@@ -118,21 +118,21 @@ from allmydata.client import (
     read_config,
 )
 
-from ..frontends.magic_folder import (
+from magic_folder.frontends.magic_folder import (
     MagicFolder,
     load_magic_folders,
     save_magic_folders,
     maybe_upgrade_magic_folders,
 )
-from ..web.magic_folder import (
+from web.magic_folder import (
     magic_folder_web_service,
 )
 
-from ..status import (
+from status import (
     status as _status,
 )
 
-from .._coverage import (
+from _coverage import (
     coverage_service,
 )
