+++ src/magic_folder/test/cli/test_magic_folder.py
@@ -50,9 +50,8 @@ from allmydata.util.eliotutil import (
 from ...frontends.magic_folder import (
     MagicFolder,
 )
-from ...scripts import (
-    magic_folder_cli,
-)
+from magic_folder import cli as magic_folder_cli
+
 from ...web.magic_folder import (
     status_for_item,
 )
