+++ src/magic_folder/test/test_magic_folder.py
@@ -45,10 +45,15 @@ from allmydata.util.eliotutil import (
     log_call_deferred,
 )
 
-from ..frontends import magic_folder
-from ..frontends.magic_folder import (
-    MagicFolder, WriteFileMixin,
+from magic_folder.frontends.magic_folder import (
+    MagicFolder,
+    WriteFileMixin,
     ConfigurationError,
+    get_inotify_module,
+    load_magic_folders,
+    maybe_upgrade_magic_folders,
+    is_new_file,
+    _upgrade_magic_folder_config,
 )
 from ..util import (
     fake_inotify,
@@ -72,7 +77,7 @@ from .cli.test_magic_folder import MagicFolderCLITestMixin
 _debug = False
 
 try:
-    magic_folder.get_inotify_module()
+    get_inotify_module()
 except NotImplementedError:
     support_missing = True
     support_message = (
@@ -162,7 +167,7 @@ class NewConfigUtilTests(SyncTestCase):
             f.write(yamlutil.safe_dump({u"magic-folders": folder_configuration}))
 
     def test_load(self):
-        folders = magic_folder.load_magic_folders(self.basedir)
+        folders = load_magic_folders(self.basedir)
         self.assertEqual(['default'], list(folders.keys()))
         self.assertEqual(folders['default'][u'umask'], 0o077)
 
@@ -178,7 +183,7 @@ class NewConfigUtilTests(SyncTestCase):
         self.folders[u"default"][u"umask"] = (0o777 & ~perm)
         self.write_magic_folder_config(self.basedir, self.folders)
 
-        magic_folder.load_magic_folders(self.basedir)
+        load_magic_folders(self.basedir)
 
         # It is created.
         self.assertTrue(
@@ -206,7 +211,7 @@ class NewConfigUtilTests(SyncTestCase):
         open(self.local_dir, "w").close()
 
         with self.assertRaises(ConfigurationError) as ctx:
-            magic_folder.load_magic_folders(self.basedir)
+            load_magic_folders(self.basedir)
         self.assertIn(
             "exists and is not a directory",
             str(ctx.exception),
@@ -224,7 +229,7 @@ class NewConfigUtilTests(SyncTestCase):
         self.write_magic_folder_config(self.basedir, self.folders)
 
         with self.assertRaises(ConfigurationError) as ctx:
-            magic_folder.load_magic_folders(self.basedir)
+            load_magic_folders(self.basedir)
         self.assertIn(
             "could not be created",
             str(ctx.exception),
@@ -233,7 +238,7 @@ class NewConfigUtilTests(SyncTestCase):
     def test_both_styles_of_config(self):
         os.unlink(join(self.basedir, u"private", u"magic_folders.yaml"))
         with self.assertRaises(Exception) as ctx:
-            magic_folder.load_magic_folders(self.basedir)
+            load_magic_folders(self.basedir)
         self.assertIn(
             "[magic_folder] is enabled but has no YAML file and no 'local.directory' option",
             str(ctx.exception)
@@ -245,7 +250,7 @@ class NewConfigUtilTests(SyncTestCase):
             f.write('----\n')
 
         with self.assertRaises(Exception) as ctx:
-            magic_folder.load_magic_folders(self.basedir)
+            load_magic_folders(self.basedir)
         self.assertIn(
             "should contain a dict",
             str(ctx.exception)
@@ -257,7 +262,7 @@ class NewConfigUtilTests(SyncTestCase):
             f.write('')
 
         with self.assertRaises(Exception) as ctx:
-            magic_folder.load_magic_folders(self.basedir)
+            load_magic_folders(self.basedir)
         self.assertIn(
             "should contain a dict",
             str(ctx.exception)
@@ -269,7 +274,7 @@ class NewConfigUtilTests(SyncTestCase):
             f.write('magic-folders: "foo"\n')
 
         with self.assertRaises(Exception) as ctx:
-            magic_folder.load_magic_folders(self.basedir)
+            load_magic_folders(self.basedir)
         self.assertIn(
             "should be a dict",
             str(ctx.exception)
@@ -290,7 +295,7 @@ class NewConfigUtilTests(SyncTestCase):
             f.write(yamlutil.safe_dump({u"magic-folders": self.folders}))
 
         with self.assertRaises(Exception) as ctx:
-            magic_folder.load_magic_folders(self.basedir)
+            load_magic_folders(self.basedir)
         self.assertIn(
             "umask must be an integer",
             str(ctx.exception)
@@ -302,7 +307,7 @@ class NewConfigUtilTests(SyncTestCase):
             f.write("magic-folders:\n  default:   foo\n")
 
         with self.assertRaises(Exception) as ctx:
-            magic_folder.load_magic_folders(self.basedir)
+            load_magic_folders(self.basedir)
         self.assertIn(
             "must itself be a dict",
             str(ctx.exception)
@@ -315,7 +320,7 @@ class NewConfigUtilTests(SyncTestCase):
             f.write(yamlutil.safe_dump({u"magic-folders": self.folders}))
 
         with self.assertRaises(Exception) as ctx:
-            magic_folder.load_magic_folders(self.basedir)
+            load_magic_folders(self.basedir)
         self.assertIn(
             "missing 'poll_interval'",
             str(ctx.exception)
@@ -370,7 +375,7 @@ class LegacyConfigUtilTests(SyncTestCase):
                 )
             )
 
-        magic_folder.load_magic_folders(self.basedir)
+        load_magic_folders(self.basedir)
 
         self.assertTrue(
             isdir(expected),
@@ -394,7 +399,7 @@ class LegacyConfigUtilTests(SyncTestCase):
             f.write("not a directory")
 
         with self.assertRaises(ConfigurationError) as ctx:
-            magic_folder.load_magic_folders(self.basedir)
+            load_magic_folders(self.basedir)
         self.assertIn(
             "is not a directory",
             str(ctx.exception)
@@ -405,7 +410,7 @@ class LegacyConfigUtilTests(SyncTestCase):
             f.write("---")
 
         with self.assertRaises(Exception) as ctx:
-            magic_folder.load_magic_folders(self.basedir)
+            load_magic_folders(self.basedir)
         self.assertIn(
             "both old-style configuration and new-style",
             str(ctx.exception)
@@ -413,7 +418,7 @@ class LegacyConfigUtilTests(SyncTestCase):
 
     def test_upgrade(self):
         # test data is created in setUp; upgrade config
-        magic_folder._upgrade_magic_folder_config(self.basedir)
+        _upgrade_magic_folder_config(self.basedir)
 
         # ensure old stuff is gone
         self.assertFalse(
@@ -435,7 +440,7 @@ class LegacyConfigUtilTests(SyncTestCase):
         self.assertFalse(config.has_option("magic_folder", "local.directory"))
 
     def test_load_legacy(self):
-        folders = magic_folder.load_magic_folders(self.basedir)
+        folders = load_magic_folders(self.basedir)
 
         self.assertEqual(['default'], list(folders.keys()))
         self.assertTrue(
@@ -449,8 +454,8 @@ class LegacyConfigUtilTests(SyncTestCase):
         )
 
     def test_load_legacy_upgrade(self):
-        magic_folder.maybe_upgrade_magic_folders(self.basedir)
-        folders = magic_folder.load_magic_folders(self.basedir)
+        maybe_upgrade_magic_folders(self.basedir)
+        folders = load_magic_folders(self.basedir)
 
         self.assertEqual(['default'], list(folders.keys()))
         # 'legacy' files should be gone
@@ -1805,10 +1810,10 @@ class SingleMagicFolderTestMixin(MagicFolderCLITestMixin, ShouldFailMixin, Reall
         #print "_wait_until_started"
         client_node = self.get_client()
         name = 'default'
-        config = magic_folder.load_magic_folders(
+        config = load_magic_folders(
             client_node.config._basedir,
         )[name]
-        self.magicfolder = magic_folder.MagicFolder.from_config(
+        self.magicfolder = MagicFolder.from_config(
             client_node,
             name,
             config,
@@ -1883,12 +1888,12 @@ class SingleMagicFolderTestMixin(MagicFolderCLITestMixin, ShouldFailMixin, Reall
         pathinfo = fileutil.get_pathinfo(path2)
         db.did_upload_version(relpath2, 0, 'URI:LIT:2', 'URI:LIT:1', 0, pathinfo)
         db_entry = db.get_db_entry(relpath2)
-        self.assertFalse(magic_folder.is_new_file(pathinfo, db_entry))
+        self.assertFalse(is_new_file(pathinfo, db_entry))
 
         different_pathinfo = fileutil.PathInfo(isdir=False, isfile=True, islink=False,
                                                exists=True, size=0, mtime_ns=pathinfo.mtime_ns,
                                                ctime_ns=pathinfo.ctime_ns)
-        self.assertTrue(magic_folder.is_new_file(different_pathinfo, db_entry))
+        self.assertTrue(is_new_file(different_pathinfo, db_entry))
 
     def _test_magicfolder_start_service(self):
         # what is this even testing?
@@ -2279,6 +2284,7 @@ class MockTestAliceBob(MagicFolderAliceBobTestMixin, AsyncTestCase):
 
     def setUp(self):
         self.inotify = fake_inotify
+        from magic_folder.frontends import magic_folder
         self.patch(magic_folder, 'get_inotify_module', lambda: self.inotify)
         return super(MockTestAliceBob, self).setUp()
 
@@ -2290,6 +2296,7 @@ class MockTest(SingleMagicFolderTestMixin, AsyncTestCase):
 
     def setUp(self):
         self.inotify = fake_inotify
+        from magic_folder.frontends import magic_folder
         self.patch(magic_folder, 'get_inotify_module', lambda: self.inotify)
         return super(MockTest, self).setUp()
 
@@ -2325,6 +2332,7 @@ class MockTest(SingleMagicFolderTestMixin, AsyncTestCase):
 
             def _not_implemented():
                 raise NotImplementedError("blah")
+            from magic_folder.frontends import magic_folder
             self.patch(magic_folder, 'get_inotify_module', _not_implemented)
             self.shouldFail(NotImplementedError, 'unsupported', 'blah',
                             MagicFolder, client, upload_dircap, '', errors_dir, magicfolderdb, 0o077, 'default')
@@ -2440,7 +2448,7 @@ class RealTest(SingleMagicFolderTestMixin, AsyncTestCase):
 
     def setUp(self):
         d = super(RealTest, self).setUp()
-        self.inotify = magic_folder.get_inotify_module()
+        self.inotify = get_inotify_module()
         return d
 
 
@@ -2451,5 +2459,5 @@ class RealTestAliceBob(MagicFolderAliceBobTestMixin, AsyncTestCase):
 
     def setUp(self):
         d = super(RealTestAliceBob, self).setUp()
-        self.inotify = magic_folder.get_inotify_module()
+        self.inotify = get_inotify_module()
         return d
