+++ src/twisted/plugins/magic_folder_dropin.py
@@ -4,7 +4,7 @@ from twisted.application.service import (
 
 magic_folder = ServiceMaker(
     "Magic-Folder for Tahoe-LAFS",
-    "magic_folder.scripts.magic_folder_cli",
+    "magic_folder.cli",
     "Tahoe-LAFS-based file synchronization",
     "magic_folder",
 )
